Add the ability to prefix the title of records created via AcceptItem messages
Some libraries want a very obvious way to see ILL records by prepending
the title with something like "INTERLIBRARY LOAN - " to make it much
more obvious than just setting the item type to ILL for example.

This commit adds a new conf option `accept_item_title_prefix` the
can be set to a string such as the one above. This string will
be prepended to the title for those records created when AcceptItem
messages are recieved.